remove ability to provide custom timing blocking interceptor and timing handler
